Added app_manager and input_handler

+++ src/color.h
@@ -4,9 +4,9 @@
 #include <stdint.h>
 
 typedef struct {
-    uint16_t r;
-    uint16_t g;
-    uint16_t b;
+    uint8_t r;
+    uint8_t g;
+    uint8_t b;
 } color_t;
 
 #endif // COLOR_H__
